to reset the water position
to reset the water position

+++ Assets/Script/NumberOfStars.cs
@@ -44,7 +44,7 @@ public class NumberOfStars : MonoBehaviour
 
     public void AddScore(float udscore)
     {
-        score = udscore;
+        score = (int)udscore;
         //HighestPercetage.instance.UpdateHighScore(score);
     }
